@@ -1,6 +1,6 @@
 package com.avadamedia.USAINUA.entity;
 
-import javax.persistence.*;
+import jakarta.persistence.*;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
