@@ -1,7 +1,7 @@
 package com.avadamedia.USAINUA.models;
 
 //import io.swagger.v3.oas.annotations.media.Schema;
-import javax.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotBlank;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Data;
